pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clementine"
version = "0.1.0"
description = "Automated web-app penetration-testing orchestrator (Project Clementine)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    # MCP client — stdio and SSE/streamable-HTTP transports
    "mcp>=1.0.0",
    # Config validation
    "pydantic>=2.0",
    "pyyaml>=6.0",
    # Async SQLite
    "aiosqlite>=0.19",
    # HTTP client used for remote MCP servers (AWS Knowledge MCP)
    "httpx>=0.27",
    # Directed-graph library for the correlation engine
    "networkx>=3.2",
    # HTML report templating
    "jinja2>=3.1",
    # CLI
    "click>=8.1",
    # AWS SDK — used to push findings to Security Hub
    "boto3>=1.34",
    # Anthropic SDK — powers AI triage and novel-chain discovery
    "anthropic>=0.43.0",
]

[project.scripts]
# Primary entry-point: `clementine run --config clementine.yaml`
clementine = "clementine.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["clementine"]
